refactor(types): split type defs in various modules & dedup
