+++ package.json
@@ -54,8 +54,8 @@
     "build:clean": "make clean",
     "check": "npm-run-all build lint depcruise test:cover",
     "check:full": "npm-run-all check test:glob test:integration",
-    "depcruise": "node ./bin/dependency-cruise --validate -- src bin bin/dependency-cruise test configs",
-    "depcruise:explain": "node ./bin/dependency-cruise --output-type err-long --validate -- src bin bin/dependency-cruise test configs",
+    "depcruise": "node ./bin/dependency-cruise --validate -- src bin bin/dependency-cruise test configs types",
+    "depcruise:explain": "node ./bin/dependency-cruise --output-type err-long --validate -- src bin bin/dependency-cruise test configs types",
     "depcruise:graph:dir": "node ./bin/dependency-cruise --validate --include-only '(^src|^[^/]+(js|json)$)' --output-type ddot bin src | dot -T png > tmp_deps.png",
     "depcruise:graph:doc": "npm-run-all depcruise:graph:doc:dot depcruise:graph:doc:html depcruise:graph:doc:png",
     "depcruise:graph:doc:dot": "node ./bin/dependency-cruise --validate --include-only '^(src|bin|bin/dependency-cruise$)' --output-type dot bin src bin/dependency-cruise --output-to tmp_graph_deps.dot",
@@ -63,15 +63,17 @@
     "depcruise:graph:doc:png": "cat tmp_graph_deps.dot | dot -Gdpi=192 -Gsplines=ortho -T png | pngquant - > doc/real-world-samples/dependency-cruiser-without-node_modules.png",
     "depcruise:graph:osage": "node ./bin/dependency-cruise --validate --include-only '(^src|^[^/]+(js|json)$)' --output-type dot bin src | osage -Gpack=16 -GpackMode=array2 -T svg > tmp_deps.svg",
     "depcruise:graph:fdp": "node ./bin/dependency-cruise --validate --include-only '^src' --output-type dot bin src | fdp -GK=0.1 -Gsplines=ortho -T svg > tmp_deps.svg",
-    "depcruise:report": "node ./bin/dependency-cruise --output-type err-html --validate -- src bin bin/dependency-cruise test configs > dependency-violations.html",
+    "depcruise:report": "node ./bin/dependency-cruise --output-type err-html --validate -- src bin bin/dependency-cruise test configs types > dependency-violations.html",
     "lint": "npm-run-all --parallel --aggregate-output lint:eslint lint:prettier:check lint:types",
     "lint:eslint": "eslint bin/dependency-cruise src test configs",
     "lint:eslint:fix": "eslint --fix bin/dependency-cruise src test configs",
     "lint:fix": "npm-run-all lint:eslint:fix lint:prettier lint:types:fix",
     "lint:prettier": "prettier --write {src,configs}/\\*\\*/\\*.{js,json} bin/\\* !**/*.template.js types/\\*.d.ts test/\\*\\*/\\*.{spec,utl}.js",
     "lint:prettier:check": "prettier --check {src,configs}/\\*\\*/\\*.{js,json} bin/\\* !**/*.template.js types/\\*.d.ts test/\\*\\*/\\*.{spec,utl}.js",
-    "lint:types": "tslint types/dependency-cruiser.d.ts",
-    "lint:types:fix": "tslint --fix types/dependency-cruiser.d.ts",
+    "lint:types": "npm-run-all lint:types:tsc lint:types:tslint",
+    "lint:types:tsc": "tsc --noEmit --strict --types --noUnusedLocals --noUnusedParameters types/dependency-cruiser.d.ts",
+    "lint:types:tslint": "tslint types/dependency-cruiser.d.ts",
+    "lint:types:fix": "tslint --fix types/*.d.ts",
     "scm:push": "run-p --aggregate-output scm:push:*",
     "scm:push:bitbucket-mirror": "run-p --aggregate-output scm:push:bitbucket-mirror:*",
     "scm:push:bitbucket-mirror:commits": "git push bitbucket-mirror",
